+++ app/src/main/java/com/sdpteam/connectout/profileList/ProfileListViewModel.java
@@ -38,7 +38,6 @@ public class ProfileListViewModel extends ViewModel {
         if(userInput != null) {
             if (option == ProfileFirebaseDataSource.ProfileOrderingOption.RATING) {
                 inputList = parseRangeInput(userInput);
-                Collections.sort(inputList);
             } else if (option == ProfileFirebaseDataSource.ProfileOrderingOption.NAME) {
                 inputList = parseNameInput(userInput);
             }
