+++ app/src/main/res/layout/activity_user_list.xml
@@ -21,7 +21,6 @@
             android:layout_height="wrap_content"
             android:src="@drawable/filter_image"
             android:drawableBottom="@drawable/filter_image"
-            android:checked = "true"
             android:textOff="Filter Off"
             android:textOn="Filter On"
             android:textSize="15dp"
