+++ app/src/main/res/layout/fragment_filter.xml
@@ -11,9 +11,7 @@
         android:id="@+id/text_filter"
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
-        android:layout_alignParentTop="true"
         android:layout_centerHorizontal="true"
-        android:layout_toLeftOf="@+id/filter_apply_button"
         android:layout_marginHorizontal="10dp"
         android:layout_marginTop="10dp"
         android:ems="10"
@@ -40,6 +38,7 @@
         android:layout_height="wrap_content"
         android:layout_alignParentBottom="true"
         android:layout_centerHorizontal="true"
+        android:layout_toLeftOf="@+id/filter_apply_button"
         android:padding="@dimen/fab_margin"
         android:orientation="horizontal">
 
@@ -74,9 +73,9 @@
     <ToggleButton
         android:id="@+id/filter_apply_button"
         android:layout_width="wrap_content"
-        android:layout_alignParentTop="true"
-        android:layout_marginTop="10dp"
-        android:layout_marginEnd="10dp"
+        android:layout_alignParentBottom="true"
+        android:layout_marginBottom="@dimen/fab_margin"
+        android:layout_marginEnd="@dimen/fab_margin"
         android:layout_centerHorizontal="true"
         android:layout_alignParentEnd="true"
         android:textOff="@string/apply_button"
